refactor(snackbar): separation files and remove function action button handler params
